+++ lib/csv2hash/railtie.rb
@@ -1,0 +1,6 @@
+require 'rails'
+
+module Csv2hash
+  class Railtie < Rails::Railtie
+  end
+end