+++ lib/generators/csv2hash/install/templates/csh2hash.rb
@@ -1,0 +1,6 @@
+Csv2hash.configure do |conf|
+  # conf.convert = false
+  # conf.true_values  = ['yes','y','t']
+  # conf.false_values = ['no','n','f']
+  # conf.nil_values   = ['nil','null']
+end